# services/rest/docs/openapi-documentation.yml
openapi: 3.1.0
info:
  title: Movie API
  version: 1.0.0
  description: REST API for authentication, movies, directors, and production companies.
servers:
  - url: http://localhost:5286
paths:
...
  /api/v1/Movies:
    get:
      summary: Get all movies (paginated)
      parameters:
        - name: page
          in: query
          required: false
          description: Page number (starts at 1)
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
            default: 1
        - name: pageSize
          in: query
          required: false
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
            default: 10
        - name: search
          in: query
          required: false
          description: Search term for movie titles
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Paginated list of movies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMovieResult'
        "204":
          description: No movies found
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
    post:
      summary: Create a new movie
      requestBody:
        description: Movie object to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieDto'
      responses:
        "201":
          description: Movie created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieHateoasDto'
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /api/v1/Movies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Movie ID
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
    get:
      summary: Get a movie by ID
      responses:
        "200":
          description: Movie found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieHateoasDto'
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseMessage'
        "500":
          description: Internal server error

    put:
      summary: Update a movie
      requestBody:
        description: Movie object to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieDto'
      responses:
        "200":
          description: Movie updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieHateoasDto'
        "400":
          description: Invalid input data or ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseMessage'
        "500":
          description: Internal server error

    delete:
      summary: Delete a movie
      responses:
        "204":
          description: Movie deleted successfully
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseMessage'
        "500":
          description: Internal server error
...
  /api/v1/ProductionCompanies:
    get:
      summary: Get all production companies (paginated)
      parameters:
        - name: page
          in: query
          required: false
          description: Page number (starts at 1)
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
            default: 1
        - name: pageSize
          in: query
          required: false
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
            default: 10
        - name: search
          in: query
          required: false
          description: Search term for production companies
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Paginated list of production companies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductionCompanyResult'
        "204":
          description: No production companies found
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /api/v1/ProductionCompanies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Production company ID
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
    get:
      summary: Get a production company by ID
      responses:
        "200":
          description: Production company found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionCompanyHateoasDto'
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBadRequest'
        "401":
          description: Unauthorized
        "404":
          description: Production company not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseMessage'
        "500":
          description: Internal server error
components:
  schemas:
    ErrorResponseBadRequest:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        traceId:
          type: string
      example:
        type: https://tools.ietf.org/html/rfc9110#section-15.5.1
        title: One or more validation errors occurred.
        status: 400
        errors:
          Password:
            - The Password field is required.
        traceId: 00-d1bd44b0408fa52dfb31a0918a49bb51
    ErrorResponseMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    MovieDto:
      type: object
      required:
        - title
        - releaseYear
        - genre
        - directorID
        - productionCompanyID
      properties:
        title:
          type: string
          maxLength: 200
        releaseYear:
          type: string
          pattern: '^\\d{4}$'
        genre:
          type: string
          maxLength: 60
        directorID:
          type: integer
          minimum: 1
          maximum: 2147483647
        productionCompanyID:
          type: integer
          minimum: 1
          maximum: 2147483647
    Link:
      type: object
      required:
        - href
        - rel
        - method
      properties:
        href:
          type: string
        rel:
          type: string
        method:
          type: string
        templated:
          type: boolean
    MovieHateoasDto:
      type: object
      required:
        - id
        - title
        - releaseYear
        - genre
        - directorID
        - productionCompanyID
        - _links
      properties:
        id:
          type: integer
        title:
          type: string
        releaseYear:
          type: string
        genre:
          type: string
        directorID:
          type: integer
        productionCompanyID:
          type: integer
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    ProductionCompanyHateoasDto:
      type: object
      required:
        - id
        - name
        - _links
      properties:
        id:
          type: integer
        name:
          type: string
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    PaginatedMovieResult:
      type: object
      required:
        - _embedded
        - page
        - pageSize
        - totalCount
        - totalPages
        - _links
      properties:
        _embedded:
          type: object
          properties:
            movies:
              type: array
              items:
                $ref: '#/components/schemas/MovieHateoasDto'
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    PaginatedProductionCompanyResult:
      type: object
      required:
        - _embedded
        - page
        - pageSize
        - totalCount
        - totalPages
        - _links
      properties:
        _embedded:
          type: object
          properties:
            productionCompanies:
              type: array
              items:
                $ref: '#/components/schemas/ProductionCompanyHateoasDto'
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer 
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'